requirements.txt
fastapi==0.115.8
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
